backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
httpx==0.27.0
groq==0.9.0
pydantic==2.6.4
email-validator==2.1.1
